refactor(core): hoist goal-step behavior into a shared core
Both engines carried a near-identical ~520-line goal step (skip-guards,
budget guard, judge/scorer resolution, activity streaming, tri-state
verdict, record persistence, feedback signal, goal chunks). The behavior
now lives once in loop/shared/steps/goal-core.ts (evaluateGoal); the two
step factories are thin glue that resolve engine dependencies (loop
params vs registry/init-data), pass a lazy message list, and project the
GoalOutcome back onto their state shape.

Adjudicated drift (unified toward the better behavior):
- Judge model-resolver returning undefined now falls back to the
  effective judgeModelId on both engines (was main-only; durable
  silently no-opped).
- The default judge scorer receives an isolated RequestContext copy on
  both engines (was main-only; durable shared the reconstructed context
  with the parent run).
- Judge activity streaming is wrapped in try/catch so a scorer stream
  failure cannot become an unhandled rejection (was durable-only).
- Chunk emission is best-effort through the injected emitChunk transport
  on both engines.

Deliberate differences kept as hooks: judge memory-thread forking
(main-only, closures cannot survive the wire), signal writer transport
(outputWriter vs pubsub), lazy message-list rehydration + serialization
(durable glue).

Verified: goal-step tests 21 passed, loop suite 1078 passed, durable
suite 794 passed with only the 3 known pre-existing evented failures.
